requirements.txt
# Q2 Pipeline Dependencies

# Core dependencies
numpy>=1.24.0
faiss-cpu>=1.7.4  # Use faiss-gpu if GPU available
tiktoken>=0.5.0  # Token counting for chunking

# API clients
mistralai>=1.0.0
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0  # For environment variable management (optional)

